+++ apps/www/src/app/(main)/(docs)/components/dialog/examples/usage.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { useDisclosure } from "@jamsr-ui/hooks";
 import {
   Button,
   Dialog,
@@ -8,38 +7,40 @@ import {
   DialogContent,
   DialogFooter,
   DialogHeader,
+  DialogTrigger,
   Repeater,
   Text,
 } from "@jamsr-ui/react";
 
 export const DialogUsage = () => {
-  const { isOpen, onClose, onOpen, setIsOpen } = useDisclosure();
   return (
-    <div>
-      <Button onClick={onOpen}>Click Me!</Button>
-      <Dialog isOpen={isOpen} onOpenChange={setIsOpen}>
-        <DialogContent>
-          <DialogHeader>Heading</DialogHeader>
-          <DialogBody className="flex flex-col gap-4">
-            <Repeater repeat={2}>
-              <Text as="p">
-                Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                Dignissimos corrupti est quos asperiores libero maiores amet non
-                obcaecati odio excepturi illo recusandae tenetur, qui earum
-                dolorem minus, quibusdam optio? Cum.
-              </Text>
-            </Repeater>
-          </DialogBody>
-          <DialogFooter>
-            <Button variant="light" onClick={onClose} color="secondary">
+    <Dialog>
+      <DialogTrigger>
+        <Button>Click Me!</Button>
+      </DialogTrigger>
+      <DialogContent>
+        <DialogHeader>Heading</DialogHeader>
+        <DialogBody className="flex flex-col gap-4">
+          <Repeater repeat={2}>
+            <Text as="p">
+              Lorem ipsum dolor sit amet consectetur adipisicing elit.
+              Dignissimos corrupti est quos asperiores libero maiores amet non
+              obcaecati odio excepturi illo recusandae tenetur, qui earum
+              dolorem minus, quibusdam optio? Cum.
+            </Text>
+          </Repeater>
+        </DialogBody>
+        <DialogFooter>
+          <DialogTrigger action="close">
+            <Button variant="light" color="secondary">
               Cancel
             </Button>
-            <Button onClick={onClose} color="primary">
-              Submit
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-    </div>
+          </DialogTrigger>
+          <DialogTrigger action="close">
+            <Button color="primary">Submit</Button>
+          </DialogTrigger>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
   );
 };