+++ apps/www/tailwind.config.ts
@@ -3,7 +3,7 @@ import { type Config } from "@jamsr-ui/tailwind-config";
 import { withJamsrUI } from "../../packages/theme/src/index";
 // import { withJamsrUI } from "../../packages/theme/src/index";
 
-const config: Config = withJamsrUI({
+const config = withJamsrUI({
   darkMode: "class",
   content: [
     "../../packages/components/*/src/**/*.{ts,tsx}",
@@ -11,5 +11,5 @@ const config: Config = withJamsrUI({
     "./src/**/*.{ts,tsx}",
     "!./node_modules",
   ],
-});
+}) satisfies Config;
 export default config;