+++ packages/components/dialog/src/dialog-context.ts
@@ -1,10 +1,11 @@
 import { createContext } from "@jamsr-ui/utils";
 import { type UseDialog } from "./use-dialog";
 
-type ContextType = ReturnType<typeof UseDialog>;
+export type DialogContextType = ReturnType<typeof UseDialog>;
 
-export const [DialogProvider, useDialogContext] = createContext<ContextType>({
-  name: "DialogContext",
-  strict: true,
-  errorMessage: "useDialogContext must be used within a Dialog",
-});
+export const [DialogProvider, useDialogContext] =
+  createContext<DialogContextType>({
+    name: "DialogContext",
+    strict: true,
+    errorMessage: "useDialogContext must be used within a Dialog",
+  });