+++ packages/components/dialog/src/index.ts
@@ -5,3 +5,5 @@ export { DialogBody, type DialogBodyProps } from "./dialog-body";
 export { DialogContent, type DialogContentProps } from "./dialog-content";
 export { DialogFooter, type DialogFooterProps } from "./dialog-footer";
 export { DialogHeader, type DialogHeaderProps } from "./dialog-header";
+export { DialogTrigger } from "./dialog-trigger";
+export { DialogState } from "./dialog-state";