+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/dialog
 
+## 0.0.103
+
+### Patch Changes
+
+- fix: dialog, otp input
+
 ## 0.0.102
 
 ### Patch Changes